refactor: merge container ports/state/identity into basics (#2908)
